fix all compiler warnings

+++ applib-dist/include/socket-lib.cpp
@@ -54,7 +54,7 @@ Connection::Connection(SOCKET sock)
 
 void Connection::send(const std::string & message) const
 {
-	::send(con_socket, message.c_str(), message.size(), 0);
+	::send(con_socket, message.c_str(), (int)message.size(), 0);
 }
 
 std::string Connection::receive() const
